Rename base EML package from eml_2_2_0 to eml
