+++ tests/test_node.py
@@ -17,8 +17,8 @@ import unittest
 
 import daiquiri
 
-from metapype.eml2_1_1 import names
-from metapype.eml2_1_1 import validate
+from metapype.eml import names
+from metapype.eml import validate
 from metapype.model.node import Node
 from metapype.model.node import Shift
